+++ tests/corpus/npbench/map_reduce/go_fast.py
@@ -3,8 +3,8 @@
 import numpy as np
 import dace as dc
 
-dc_float = dc.float32
-dc_complex_float = dc.complex64
+dc_float = dc.float64
+dc_complex_float = dc.complex128
 
 SIZES = {'N': 2000}
 INPUT_ARGS = ('N', )
@@ -15,7 +15,7 @@ OUTPUT_ARGS = ('out', )
 N = dc.symbol('N', dtype=dc.int64)
 
 
-def initialize(N, datatype=np.float32):
+def initialize(N, datatype=np.float64):
     from numpy.random import default_rng
     rng = default_rng(42)
     x = rng.random((N, N), dtype=datatype)
