+++ tests/corpus/npbench/ml/mlp.py
@@ -3,8 +3,8 @@
 import numpy as np
 import dace as dc
 
-dc_float = dc.float32
-dc_complex_float = dc.complex64
+dc_float = dc.float64
+dc_complex_float = dc.complex128
 
 SIZES = {'C_in': 3, 'N': 8, 'S0': 30000, 'S1': 2000, 'S2': 2000}
 INPUT_ARGS = ('C_in', 'N', 'S0', 'S1', 'S2')
@@ -15,7 +15,7 @@ OUTPUT_ARGS = ('out', )
 C_in, N, S0, S1, S2, N1, N2 = (dc.symbol(s, dtype=dc.int64) for s in ('C_in', 'N', 'S0', 'S1', 'S2', 'N1', 'N2'))
 
 
-def initialize(C_in, N, S0, S1, S2, datatype=np.float32):
+def initialize(C_in, N, S0, S1, S2, datatype=np.float64):
     from numpy.random import default_rng
     rng = default_rng(42)
     mlp_sizes = [S0, S1, S2]
