+++ tests/corpus/npbench/ml/softmax.py
@@ -3,8 +3,8 @@
 import numpy as np
 import dace as dc
 
-dc_float = dc.float32
-dc_complex_float = dc.complex64
+dc_float = dc.float64
+dc_complex_float = dc.complex128
 
 SIZES = {'N': 16, 'H': 16, 'SM': 128}
 INPUT_ARGS = ('N', 'H', 'SM')
@@ -15,7 +15,7 @@ OUTPUT_ARGS = ('out', )
 N, H, SM = (dc.symbol(s, dc.int64) for s in ('N', 'H', 'SM'))
 
 
-def initialize(N, H, SM, datatype=np.float32):
+def initialize(N, H, SM, datatype=np.float64):
     from numpy.random import default_rng
     rng = default_rng(42)
     x = rng.random((N, H, SM, SM), dtype=datatype)
